+++ iqshell/list_bucket.go
@@ -58,11 +58,17 @@ func errorWarning(marker string, err error) {
 *@return listError
  */
 func (m *BucketManager) ListFiles(bucket, prefix, marker, listResultFile string) (retErr error) {
-	return m.ListBucket2(bucket, prefix, marker, listResultFile, "", time.Time{}, time.Time{}, nil, 20, false)
+	return m.ListBucket2(bucket, prefix, marker, listResultFile, "", time.Time{}, time.Time{}, nil, 20, false, false)
 }
 
-func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, delimiter string, startDate, endDate time.Time, suffixes []string, maxRetry int, appendMode bool) (retErr error) {
+func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, delimiter string, startDate, endDate time.Time, suffixes []string, maxRetry int, appendMode bool, readable bool) (retErr error) {
 	lastMarker := marker
+
+	defer func(lastMarker string) {
+		if lastMarker != "" {
+			fmt.Fprintf(os.Stderr, "Marker: %s\n", lastMarker)
+		}
+	}(lastMarker)
 
 	sigChan := make(chan os.Signal, 1)
 	ctx, cancel := context.WithCancel(context.Background())
@@ -74,6 +80,9 @@ func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, deli
 		<-sigChan
 		cancel()
 		maxRetry = 0
+
+		fmt.Printf("\nMarker: %s\n", lastMarker)
+		os.Exit(1)
 	}()
 
 	var listResultFh *os.File
@@ -128,6 +137,7 @@ func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, deli
 			time.Sleep(1)
 			continue
 		}
+		var fsizeValue interface{}
 
 		for listItem := range entries {
 			if listItem.Marker != lastMarker {
@@ -136,9 +146,14 @@ func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, deli
 			if listItem.Item.IsEmpty() {
 				continue
 			}
+			if readable {
+				fsizeValue = BytesToReadable(listItem.Item.Fsize)
+			} else {
+				fsizeValue = listItem.Item.Fsize
+			}
 			if notfilterSuffix && notfilterTime {
-				lineData := fmt.Sprintf("%s\t%d\t%s\t%d\t%s\t%d\t%s\r\n",
-					listItem.Item.Key, listItem.Item.Fsize, listItem.Item.Hash,
+				lineData := fmt.Sprintf("%s\t%v\t%s\t%d\t%s\t%d\t%s\r\n",
+					listItem.Item.Key, fsizeValue, listItem.Item.Hash,
 					listItem.Item.PutTime, listItem.Item.MimeType, listItem.Item.Type, listItem.Item.EndUser)
 				_, wErr := bWriter.WriteString(lineData)
 				if wErr != nil {
@@ -160,8 +175,8 @@ func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, deli
 				}
 
 				if hasSuffix && putTimeValid {
-					lineData := fmt.Sprintf("%s\t%d\t%s\t%d\t%s\t%d\t%s\r\n",
-						listItem.Item.Key, listItem.Item.Fsize, listItem.Item.Hash,
+					lineData := fmt.Sprintf("%s\t%v\t%s\t%d\t%s\t%d\t%s\r\n",
+						listItem.Item.Key, fsizeValue, listItem.Item.Hash,
 						listItem.Item.PutTime, listItem.Item.MimeType, listItem.Item.Type, listItem.Item.EndUser)
 					_, wErr := bWriter.WriteString(lineData)
 					if wErr != nil {
@@ -186,8 +201,5 @@ func (m *BucketManager) ListBucket2(bucket, prefix, marker, listResultFile, deli
 		}
 	}
 
-	if lastMarker != "" {
-		fmt.Println("Marker: ", lastMarker)
-	}
 	return
 }